+++ docs/conf.py
@@ -17,6 +17,8 @@ import sys
 import subprocess
 from typing import Any, Dict, List
 
+from sphinx.util import logging
+
 sys.path.insert(0, os.path.abspath("."))
 sys.path.insert(0, os.path.abspath("../"))
 from pystac.version import __version__, STACVersion  # noqa:E402
@@ -231,8 +233,17 @@ epub_exclude_files = ["search.html"]
 
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3", None),
+    "dateutil": ("https://dateutil.readthedocs.io/en/stable", None),
 }
 
 # -- Substutition variables
 
 rst_epilog = f".. |stac_version| replace:: {STACVersion.DEFAULT_STAC_VERSION}"
+
+# -- Suppress warnings from the extlinks extension
+# We do this to avoid warnings like the following in our Jupyter notebook tutorials
+# where we do not want to use Sphinx constructs:
+# WARNING: hardcoded link 'https://github.com/stac-extensions/eo' could be replaced
+# by an extlink (try using ':stac-ext:`eo`' instead)
+linklogger = logging.getLogger("sphinx.ext.extlinks")
+linklogger.setLevel(40)  # Ignore messages less severe than ERROR