+++ pystac/extensions/table.py
@@ -1,7 +1,4 @@
-"""Implements the Table extension
-
-https://github.com/stac-extensions/table
-"""
+"""Implements the :stac-ext:`Table Extension <table>.`"""
 from typing import Any, Dict, Generic, List, Optional, TypeVar, Union, cast
 
 import pystac